Refactor theming types & interfaces
