+++ packages/theming/.size-snapshot.json
@@ -1,13 +1,13 @@
 {
   "index.cjs.js": {
-    "bundled": 22763,
-    "minified": 14425,
-    "gzipped": 5318
+    "bundled": 23081,
+    "minified": 14712,
+    "gzipped": 5405
   },
   "index.esm.js": {
-    "bundled": 21908,
-    "minified": 13648,
-    "gzipped": 5193,
+    "bundled": 22212,
+    "minified": 13925,
+    "gzipped": 5277,
     "treeshaked": {
       "rollup": {
         "code": 3408,
